+++ src/realm/table_view.cpp
@@ -681,12 +681,12 @@ void TableViewBase::sync_distinct_view(size_t column)
 
 void TableViewBase::distinct(size_t column)
 {
-    distinct(std::vector<size_t> { column });
+    distinct(std::vector<LinkChain> { LinkChain(column) });
 }
 
 /// Remove rows that are duplicated with respect to the column set passed as argument.
 /// Will keep original sorting order so that you can both have a distinct and sorted view.
-void TableViewBase::distinct(std::vector<size_t> columns)
+void TableViewBase::distinct(std::vector<LinkChain> columns)
 {
     m_distinct_columns.clear();
     const_cast<TableViewBase*>(this)->do_sync();
@@ -714,7 +714,7 @@ void TableViewBase::distinct(std::vector<size_t> columns)
     std::vector<const ColumnBase*> m_columns;
     m_columns.resize(columns.size());
     for (size_t i = 0; i < columns.size(); i++) {
-        m_columns[i] = &m_table->get_column_base(m_distinct_columns[i]);
+        m_columns[i] = &m_table->get_column_base(m_distinct_columns[i][0]);
     }
 
     // Step 4: Build a list of all duplicated rows that need to be removed
@@ -755,15 +755,16 @@ void TableViewBase::distinct(std::vector<size_t> columns)
 // Sort according to one column
 void TableViewBase::sort(size_t column, bool ascending)
 {
-    std::vector<size_t> c;
-    std::vector<bool> a;
-    c.push_back(column);
-    a.push_back(ascending);
-    sort(c, a);
+    sort({LinkChain(column)}, std::vector<bool>{ascending});
+}
+
+void TableViewBase::sort(LinkChain chain, bool ascending)
+{
+    sort({chain}, std::vector<bool>{ascending});
 }
 
 // Sort according to multiple columns, user specified order on each column
-void TableViewBase::sort(std::vector<size_t> columns, std::vector<bool> ascending)
+void TableViewBase::sort(std::vector<LinkChain> columns, std::vector<bool> ascending)
 {
     REALM_ASSERT(columns.size() == ascending.size());
     m_auto_sort = true;