+++ src/realm/views.cpp
@@ -16,11 +16,86 @@
  *
  **************************************************************************/
 
-#include <realm/table.hpp>
-#include <realm/impl/sequential_getter.hpp>
 #include <realm/views.hpp>
 
+#include <realm/column_link.hpp>
+#include <realm/table.hpp>
+
 using namespace realm;
+
+LinkChain::LinkChain(size_t single_index)
+    : m_column_indices{single_index}
+{}
+
+LinkChain::LinkChain(std::vector<size_t> chain)
+    : m_column_indices(std::move(chain))
+{
+    REALM_ASSERT(m_column_indices.size() >= 1);
+}
+
+const ColumnBase& LinkChain::init(const ColumnBase* cb, IntegerColumn* row_indexes)
+{
+    REALM_ASSERT(cb != nullptr);
+    REALM_ASSERT(row_indexes != nullptr);
+
+    typedef _impl::TableFriend tf;
+
+    if (m_column_indices.size() > 1) {
+        size_t num_rows = row_indexes->size();
+
+        m_link_translator = std::make_shared<NullableVector>(num_rows);
+
+        std::vector<const LinkColumn*> link_cols;
+        const Table* linked_table = nullptr;
+        const ColumnBase* next_col = cb;
+        for (size_t link_ndx = 0; link_ndx < m_column_indices.size() - 1; link_ndx++) {
+            const LinkColumn* link_col = dynamic_cast<const LinkColumn*>(next_col);
+            // Only last column in link chain is allowed to be non-link
+            if (!link_col) {
+                throw LogicError(LogicError::type_mismatch);
+            }
+            link_cols.push_back(link_col);
+            linked_table = &link_col->get_target_table();
+            next_col = &tf::get_column(*linked_table, m_column_indices[link_ndx + 1]);
+        }
+
+        for (size_t row_ndx = 0; row_ndx < num_rows; row_ndx++) {
+            size_t translated_index = row_indexes->get(row_ndx);
+            bool set_null = false;
+            for (const LinkColumn* link_col : link_cols) {
+                if (link_col->is_null(translated_index)) {
+                    set_null = true;
+                    break;
+                }
+                else {
+                    translated_index = link_col->get_link(translated_index);
+                }
+            }
+            REALM_ASSERT_EX(row_ndx < m_link_translator->size(), row_ndx, m_link_translator->size());
+            if (set_null) {
+                (*m_link_translator)[row_ndx] = {};
+            }
+            else {
+                (*m_link_translator)[row_ndx] = { translated_index };
+            }
+        }
+        REALM_ASSERT(linked_table);
+        const ColumnBase& last_col_in_chain = tf::get_column(*linked_table, m_column_indices.back());
+        return last_col_in_chain;
+    }
+    return *cb; // no link chain, return original column
+}
+
+util::Optional<size_t> LinkChain::translate(size_t index) const
+{
+    util::Optional<size_t> value(index);
+    if (m_link_translator) {
+        REALM_ASSERT_EX(index < m_link_translator->size(), index, m_link_translator->size());
+        value = (*m_link_translator)[index];
+    }
+    return value;
+}
+
 
 // Re-sort view according to last used criterias
 void RowIndexes::sort(Sorter& sorting_predicate)
@@ -29,7 +104,7 @@ void RowIndexes::sort(Sorter& sorting_predicate)
     if (sz == 0)
         return;
 
-    std::vector<size_t> v;
+    std::vector<IndexPair> v;
     v.reserve(sz);
     // always put any detached refs at the end of the sort
     // FIXME: reconsider if this is the right thing to do
@@ -39,16 +114,17 @@ void RowIndexes::sort(Sorter& sorting_predicate)
     for (size_t t = 0; t < sz; t++) {
         int64_t ndx = m_row_indexes.get(t);
         if (ndx != detached_ref) {
-            v.push_back(ndx);
+            v.push_back(IndexPair(ndx, t));
         }
         else
             ++detached_ref_count;
     }
     sorting_predicate.init(this);
     std::stable_sort(v.begin(), v.end(), sorting_predicate);
+    sorting_predicate.cleanup();
     m_row_indexes.clear();
     for (size_t t = 0; t < sz - detached_ref_count; t++)
-        m_row_indexes.add(v[t]);
+        m_row_indexes.add(v[t].index_in_column);
     for (size_t t = 0; t < detached_ref_count; ++t)
         m_row_indexes.add(-1);
 }